+++ env_algorithm/BME.h
@@ -19,8 +19,8 @@ class BME
         int CAN_ID;
 
         // print_data
-        uint8_t data[BME280_num_of_variables][8][8];
-        int time_data[8][8];
+        byte data[BME280_num_of_variables][8][8];
+        byte time_data[8][8];
 
         // debug
         String name = "BME280";
