+++ env_algorithm/CO2.h
@@ -21,8 +21,8 @@ class CO2
         int CAN_ID;
 
         // data
-        uint8_t data[CO2_num_of_variables][8][8];
-        int     time_data[8][8];
+        byte data[CO2_num_of_variables][8][8];
+        byte     time_data[8][8];
 
         // debug
         String name = "CO2";
