+++ env_algorithm/L0.h
@@ -19,8 +19,8 @@ class L0
         int CAN_ID;
 
         // data
-        uint8_t data[L0_num_of_variables][8][8];
-        int     time_data[8][8];
+        byte data[L0_num_of_variables][8][8];
+        byte     time_data[8][8];
 
         // debug
         String  name = "L0";
