+++ env_algorithm/global.h
@@ -1,12 +1,14 @@
 
 #include <stdint.h>
 
+#include <Arduino.h>
+
 #define debug
 #define serial_debug            Serial
 #define serial_debug_speed      115200
 
-void update_16bit(uint8_t data[][8][8], 
-                  int time_data[8][8],
+void update_16bit(byte data[][8][8], 
+                  byte time_data[8][8],
                   int var_id, 
                   int& col, 
                   int& row, 
@@ -14,8 +16,8 @@ void update_16bit(uint8_t data[][8][8],
                   int number_of_variables,
                   int value);
 
-void print_debug(uint8_t data[][8][8], 
-                 int time_data[8][8],
+void print_debug(byte data[][8][8], 
+                 byte time_data[8][8],
                  int number_of_variables,
                  int col,
                  int row,
